Deleted pladipus specific classes, simplified peptideshaker and searchgui steps
