+++ reanalysis-pipeline/pom.xml
@@ -63,10 +63,18 @@
             <artifactId>commons-math</artifactId>
             <version>${common-math.version}</version>
         </dependency>
-
-
-
-
+        <dependency>
+            <groupId>uk.ac.ebi.pride.utilities</groupId>
+            <artifactId>ms-data-core-api</artifactId>
+            <version>2.0.15-SNAPSHOT</version>
+            <type>jar</type>
+        </dependency>
+        <dependency>
+            <groupId>com.compomics</groupId>
+            <artifactId>utilities</artifactId>
+            <version>4.11.17</version>
+            <type>jar</type>
+        </dependency>
     </dependencies>
 
     <build>
