+++ reanalysis-pipeline/src/main/java/uk/ac/eb/pride/cluster/reanalysis/control/util/CommandLineArgumentPrinter.java
@@ -1,0 +1,26 @@
+package uk.ac.eb.pride.cluster.reanalysis.control.util;
+
+import uk.ac.eb.pride.cluster.reanalysis.model.enums.AllowedPeptideShakerParams;
+import uk.ac.eb.pride.cluster.reanalysis.model.enums.AllowedSearchGUIParams;
+
+/**
+ *
+ * @author Kenneth Verheggen <kenneth.verheggen@ugent.be>
+ */
+public class CommandLineArgumentPrinter {
+
+    public static void main(String[] args) {
+        System.out.println("TO RUN A SEARCHGUI STEP : ");
+        for (AllowedSearchGUIParams aParam : AllowedSearchGUIParams.values()) {
+            if (aParam.isMandatory()) {
+                System.out.println(aParam.id + "\t" + aParam.getDescription());
+            }
+        }
+        System.out.println("ADDITIONALLY TO RUN A PEPTIDESHAKER STEP : ");
+        for (AllowedPeptideShakerParams aParam : AllowedPeptideShakerParams.values()) {
+            if (aParam.isMandatory()) {
+                System.out.println(aParam.id + "\t" + aParam.getDescription());
+            }
+        }
+    }
+}
