+++ reanalysis-pipeline/src/main/java/uk/ac/eb/pride/cluster/reanalysis/model/processing/processsteps/SearchSetupStep.java
@@ -1,4 +1,4 @@
-package uk.ac.eb.pride.cluster.reanalysis.processsteps;
+package uk.ac.eb.pride.cluster.reanalysis.model.processing.processsteps;
 
 import com.compomics.util.experiment.biology.taxonomy.SpeciesFactory;
 import com.compomics.util.experiment.identification.identification_parameters.SearchParameters;
@@ -9,13 +9,14 @@ import uk.ac.eb.pride.cluster.reanalysis.control.util.ZipUtils;
 import uk.ac.eb.pride.cluster.reanalysis.model.exception.ProcessingException;
 import uk.ac.eb.pride.cluster.reanalysis.model.exception.UnspecifiedException;
 import uk.ac.eb.pride.cluster.reanalysis.model.processing.ProcessingStep;
-import uk.ac.eb.pride.cluster.reanalysis.util.PladipusFileDownloadingService;
+import uk.ac.eb.pride.cluster.reanalysis.control.util.PladipusFileDownloadingService;
 
 import javax.xml.stream.XMLStreamException;
 import java.io.File;
 import java.io.FileNotFoundException;
 import java.io.IOException;
 import java.net.URISyntaxException;
+import uk.ac.eb.pride.cluster.reanalysis.model.GlobalProcessingProperties;
 
 /**
  *
@@ -24,21 +25,136 @@ import java.net.URISyntaxException;
 public class SearchSetupStep extends ProcessingStep {
 
     /**
+     * The logging instance
+     */
+    private static final Logger LOGGER = Logger.getLogger(SearchSetupStep.class);
+
+    /**
      * the temp folder for the entire processing
      */
     private final File tempResources;
+
+    /**
+     * the fasta folder for processing. This is required to ensure peptideshaker
+     * does not rebuild protein trees (computationally heavy task that takes a
+     * lot of time, but should only be done once)
+     */
     private final File fasta_repo;
-    private static final Logger LOGGER = Logger.getLogger(SearchSetupStep.class);
 
     public SearchSetupStep() {
-        tempResources = new File(System.getProperty("user.home") + "/pladipus/temp/search/resources");
+        tempResources = GlobalProcessingProperties.TEMP_FOLDER;
         tempResources.getParentFile().mkdirs();
-        fasta_repo = new File(System.getProperty("user.home") + "/pladipus/fasta");
+        fasta_repo = GlobalProcessingProperties.FASTA_REPOSITORY_FOLDER;
         fasta_repo.mkdirs();
     }
 
     @Override
     public boolean doAction() throws ProcessingException {
+        cleanUp();
+        try {
+            initialiseInputFiles();
+        } catch (Exception ex) {
+            throw new ProcessingException(ex);
+        }
+        return true;
+    }
+
+    private void initialiseInputFiles() throws Exception {
+        //original
+        String inputPath = parameters.get("spectrum_files");
+        String fastaPath = parameters.get("fasta_file");
+
+        if (!inputPath.equalsIgnoreCase(tempResources.getAbsolutePath())) {
+
+            if (inputPath.toLowerCase().endsWith(".mgf")
+                    || inputPath.toLowerCase().endsWith(".mgf.zip")) {
+
+                //move the input file to the temporary file using the pladipus file downloading service (it should be able 
+                //to handle uri's as well that way
+                File downloadFile = PladipusFileDownloadingService.downloadFile(inputPath, tempResources);
+                downloadFile.deleteOnExit();
+                String inputFile = downloadFile.getAbsolutePath();
+                //if it is zipped, unzip it...
+                if (inputPath.toLowerCase().endsWith(".zip")) {
+                    LOGGER.info("Unzipping input");
+                    ZipUtils.unzipArchive(new File(inputFile), tempResources);
+                }
+                //set the mgf as a spectrum file
+                parameters.put("spectrum_files", inputFile.replace(".zip", ""));
+            }
+        }
+        LoadFasta(fastaPath);
+        parameters.put("search_setup_done", "true");
+    }
+
+    public void LoadFasta(String fastaPath) throws FileNotFoundException, IOException, ClassNotFoundException, XMLStreamException, URISyntaxException, UnspecifiedException {
+        LOGGER.info("Updating the provided search parameters with the requested fasta");
+        String paramPath = parameters.get("id_params");
+        //generate a repo folder for fasta files...
+        //clear the repository save for the current fasta (temporary solution) 
+        String fastaName = new File(fastaPath).getName();
+
+        //check if this fasta was used before, so we can skip peptideshaker's 
+        //protein tree building
+        boolean fastaAlreadyExists = false;
+        File fastaFile = null;
+        for (File aFasta : fasta_repo.listFiles()) {
+            if (aFasta.getName().equalsIgnoreCase(fastaName)) {
+                fastaFile = aFasta;
+                fastaAlreadyExists = true;
+                break;
+            }
+        }
+        //if it's not there, create it there
+        if (!fastaAlreadyExists) {
+            fastaFile = PladipusFileDownloadingService.downloadFile(fastaPath, fasta_repo, fastaName);
+        } else {
+            //do we want to delete all other fastas?
+            /*
+            for (File aFile : fasta_repo.listFiles()) {
+                if (!aFile.getName().toLowerCase().contains(fastaName.toLowerCase())) {
+                    aFile.delete();
+                }
+            }*/
+        }
+
+        //if all has completed properly, there should be a fasta file here now...
+        if (fastaFile != null) {
+            parameters.put("fasta_file", fastaFile.getAbsolutePath());
+
+            //get and update parameters
+            File paramFile;
+            if (!paramPath.contains(tempResources.getAbsolutePath())) {
+                paramFile = PladipusFileDownloadingService.downloadFile(paramPath, tempResources);
+                parameters.put("id_params", paramFile.getAbsolutePath());
+            }
+            paramFile = new File(parameters.get("id_params"));
+            SearchParameters sparameters = SearchParameters.getIdentificationParameters(paramFile);
+            IdentificationParameters updatedIdentificationParameters = updateAlgorithmSettings(sparameters, fastaFile);
+            IdentificationParameters.saveIdentificationParameters(updatedIdentificationParameters, paramFile);
+
+            //output
+            File outputFolder = new File(parameters.get("output_folder"));
+            outputFolder.mkdirs();
+            parameters.put("output_folder", outputFolder.getAbsolutePath());
+        } else {
+            throw new FileNotFoundException("Fasta file was not found !");
+        }
+    }
+
+    @Override
+    public String getDescription() {
+        return "Initialisation of the search process";
+    }
+
+    public IdentificationParameters updateAlgorithmSettings(SearchParameters searchParameters, File fasta) throws IOException, XMLStreamException, URISyntaxException, UnspecifiedException {
+        searchParameters.setFastaFile(fasta);
+        SpeciesFactory.getInstance().initiate(new SearchGUIStep().getJar().getParentFile().getAbsolutePath());
+        IdentificationParameters temp = new IdentificationParameters(searchParameters);
+        return temp;
+    }
+
+    private void cleanUp() throws ProcessingException {
         if (!parameters.containsKey("skip_cleaning") && tempResources.exists()) {
             LOGGER.info("Cleaning up resources : ");
             for (File aFile : tempResources.listFiles()) {
@@ -58,105 +174,6 @@ public class SearchSetupStep extends ProcessingStep {
         } else {
             tempResources.mkdirs();
         }
-        try {
-            initialiseInputFiles();
-        } catch (Exception ex) {
-            throw new ProcessingException(ex);
-        }
-        return true;
-    }
-
-
-    private void initialiseInputFiles() throws Exception {
-        //original
-        String inputPath = parameters.get("spectrum_files");
-        String fastaPath = parameters.get("fasta_file");
-
-        //fix for older files that lack identification parameters
-        if (!inputPath.equalsIgnoreCase(tempResources.getAbsolutePath())) {
-            if (inputPath.toLowerCase().endsWith("raw.zip") || inputPath.toLowerCase().endsWith(".raw") || inputPath.toLowerCase().endsWith(".mgf") || inputPath.toLowerCase().endsWith(".mgf.zip")) {
-                System.out.println("RAW file or mgf was identified");
-                File downloadFile = PladipusFileDownloadingService.downloadFile(inputPath, tempResources);
-                downloadFile.deleteOnExit();
-                String inputFile = downloadFile.getAbsolutePath();
-
-                if (inputPath.toLowerCase().endsWith(".zip")) {
-                    LOGGER.debug("Unzipping input");
-                    ZipUtils.unzipArchive(new File(inputFile), tempResources);
-                    LOGGER.debug("Done unzipping...");
-                }
-                parameters.put("spectrum_files", inputFile.replace(".zip", ""));
-            } else {
-                parameters.put("spectrum_files", PladipusFileDownloadingService.downloadFolder(inputPath, tempResources).getAbsolutePath());
-            }
-        }
-        LOGGER.info("Got input files " + parameters.get("spectrum_files"));
-        LoadFasta(fastaPath);
-    }
-
-    public void LoadFasta(String fastaPath) throws Exception {
-        String paramPath = parameters.get("id_params");
-        //generate a repo folder for fasta files...
-        //clear the repository save for the current fasta (temporary solution)
-        //TODO refactor that it deletes complete runs only
-        String fastaName = new File(fastaPath).getName();
-        boolean fastaAlreadyExists = false;
-        File fastaFile = null;
-        for (File aFasta : fasta_repo.listFiles()) {
-            if (aFasta.getName().equalsIgnoreCase(fastaName)) {
-                fastaFile = aFasta;
-                fastaAlreadyExists = true;
-                break;
-            }
-        }
-        if (!fastaAlreadyExists) {
-            fastaFile = PladipusFileDownloadingService.downloadFile(fastaPath, fasta_repo, fastaName);
-        } else {
-            for (File aFile : fasta_repo.listFiles()) {
-                if (!aFile.getName().toLowerCase().contains(fastaName.toLowerCase())) {
-                    aFile.delete();
-                }
-            }
-        }
-
-        if (fastaFile != null) {
-            parameters.put("fasta_file", fastaFile.getAbsolutePath());
-            LOGGER.debug("Got fasta file " + parameters.get("fasta_file"));
-            //get and update parameters
-            File paramFile;
-            if (!paramPath.contains(tempResources.getAbsolutePath())) {
-                paramFile = PladipusFileDownloadingService.downloadFile(paramPath, tempResources);
-                parameters.put("id_params", paramFile.getAbsolutePath());
-            }
-            paramFile = new File(parameters.get("id_params"));
-            SearchParameters sparameters = SearchParameters.getIdentificationParameters(paramFile);
-            IdentificationParameters updatedIdentificationParameters = updateAlgorithmSettings(sparameters, fastaFile);
-            IdentificationParameters.saveIdentificationParameters(updatedIdentificationParameters, paramFile);
-
-            LOGGER.info("Got identification parameters " + parameters.get("id_params"));
-
-            //output
-            File outputFolder = new File(parameters.get("output_folder"));
-            outputFolder.mkdirs();
-            LOGGER.debug("Got output folder " + parameters.get("output_folder"));
-            parameters.put("output_folder", outputFolder.getAbsolutePath());
-        } else {
-            throw new FileNotFoundException("Fasta file was not found !");
-        }
-    }
-
-    @Override
-    public String getDescription() {
-        return "Initialisation of the search process";
-    }
-
-    public IdentificationParameters updateAlgorithmSettings(SearchParameters searchParameters, File fasta) throws IOException, XMLStreamException, URISyntaxException, UnspecifiedException {
-        System.out.println("Updating the algorithm settings and setting the fasta file...");
-        searchParameters.setFastaFile(fasta);
-        SpeciesFactory.getInstance().initiate(new SearchGUIStep().getJar().getParentFile().getAbsolutePath());
-
-        IdentificationParameters temp = new IdentificationParameters(searchParameters);
-        return temp;
     }
 
     public static void main(String[] args) {
